Move `confirmExports` to shared internal-test-helpers.
